AI_Math_tutor/requirements.txt
# Core
streamlit>=1.32.0
python-dotenv>=1.0.0

# LLM & RAG
google-generativeai>=0.4.0
langchain>=0.1.12
langchain-google-genai>=1.0.1
langchain-community>=0.0.28
chromadb>=0.4.22

# Agents / Symbolic Math
sympy>=1.12

# OCR
easyocr>=1.7.1
pillow>=10.2.0
opencv-python-headless>=4.9.0

# Speech to Text
openai-whisper>=20231117

# Memory
sqlalchemy>=2.0.28

# Utilities
numpy>=1.26.4
requests>=2.31.0
